% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend-helpers.R
\name{econodist_legend_grob}
\alias{econodist_legend_grob}
\title{Create a legend grob that can be used with econodist charts}
\usage{
econodist_legend_grob(family = "EconSansCndLig", label_size = 10,
  tenth_lab = "10th percentile", tenth_col = econ_tenth,
  med_lab = "Median", med_col = econ_median,
  ninetieth_lab = "90th percentile", ninetieth_col = econ_ninetieth,
  label_col = econ_text_col)
}
\arguments{
\item{family}{font family}

\item{label_size}{size of legend text}

\item{tenth_lab, tenth_col}{label & color for the tenth bar}

\item{med_lab, med_col}{label & color for the median point}

\item{ninetieth_lab, ninetieth_col}{lanbel & color for the ninetieth bar}

\item{label_col}{color of the legend text}
}
\description{
Create a legend grob that can be used with econodist charts
}
\seealso{
Other Econodist legend helpers: \code{\link{add_econodist_legend}},
  \code{\link{left_align}}
}
\concept{Econodist legend helpers}
