% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-econodist.R, R/stat-econodist.R
\docType{data}
\name{geom_econodist}
\alias{geom_econodist}
\alias{GeomEconodist}
\alias{stat_econodist}
\alias{StatEconodist}
\title{Econodist geom / stat}
\format{An object of class \code{GeomEconodist} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.}
\usage{
geom_econodist(mapping = NULL, data = NULL, stat = "econodist",
  position = "dodge2", tenth_col = econ_tenth,
  median_col = econ_median, ninetieth_col = econ_ninetieth,
  median_point_size = NULL, endcap_adjust = 1.5, ..., na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)

GeomEconodist

stat_econodist(mapping = NULL, data = NULL, geom = "econodist",
  position = "dodge2", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

StatEconodist
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()} or
\code{aes_()}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{ggplot()}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{fortify()} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{stat}{ggplot2 stat to use}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{tenth_col, median_col, ninetieth_col, median_point_size}{colors for geom components}

\item{endcap_adjust}{multipler to make endcaps wider/thinner}

\item{...}{other arguments passed on to \code{layer()}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{borders()}.}

\item{geom}{ggplot2 geom to use}
}
\description{
Like \code{\link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot()}} you can either pass in pre-computed
values for "ymin", "median", and "ymax" or a single y column
which will then use \code{\link[=stat_econodist]{stat_econodist()}} to compute the needed
statistics.
}
\examples{
ggplot(mammogram_costs, aes(x = city)) +
  geom_econodist(
    aes(
      ymin = tenth,
      median = median,
      ymax = ninetieth
   ),
    stat = "identity",
  ) +
  scale_y_continuous(
    expand = c(0,0),
    position = "right",
    limits = range(0, 800),
    label = scales::comma
  ) +
  coord_flip() +
  labs(
    x = NULL, y = NULL
  )
}
\keyword{datasets}
