% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggeconodist-package.R
\docType{package}
\name{ggeconodist}
\alias{ggeconodist}
\alias{ggeconodist-package}
\title{Create Diminutive Distribution Charts}
\description{
'The Economist' has a unique boxplot aesthetic for
communicating distrribution characteristics. Tools are provided
to create similar charts in 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/ggeconodist}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/ggeconodist/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
