% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend-helpers.R
\name{left_align}
\alias{left_align}
\title{Helper to flush ggplot2 plot components to the left}
\usage{
left_align(gg, components)
}
\arguments{
\item{gg}{ggplot2 plot}

\item{components}{ggplot2 named gtable components to operate on}
}
\description{
Stolen from the BBC (don't tell Scotland Yard)
}
\seealso{
Other Econodist legend helpers: \code{\link{add_econodist_legend}},
  \code{\link{econodist_legend_grob}}
}
\concept{Econodist legend helpers}
