% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{mammogram_costs}
\alias{mammogram_costs}
\title{Cost of a mammogram in various U.S. Citites (2016, USD)}
\format{An object of class \code{data.frame} with 11 rows and 5 columns.}
\usage{
mammogram_costs
}
\description{
Cost of a mammogram in various U.S. Citites (2016, USD)
}
\keyword{datasets}
