% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-econ.R
\name{theme_econodist}
\alias{theme_econodist}
\title{A more current Economist-style ggplot2 theme}
\usage{
theme_econodist(econ_text_col = "#3b454a",
  econ_plot_bg_col = "#d7e6ee", econ_grid_col = "#bbcad2",
  econ_font = "EconSansCndReg", light_font = "EconSansCndLig",
  bold_font = "EconSansCndBol")
}
\arguments{
\item{econ_text_col}{color for text elements}

\item{econ_plot_bg_col}{plot background color}

\item{econ_grid_col}{plot grid color}

\item{econ_font}{core plot font}

\item{light_font}{light font used in various polaces}

\item{bold_font}{bold font used in various places}
}
\description{
A more current Economist-style ggplot2 theme
}
\note{
You \emph{need} their fonts installed. You can get them from
\href{https://github.com/economist-components/component-typography}{here}
}
