% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend-helpers.R
\name{add_econodist_legend}
\alias{add_econodist_legend}
\title{Helper utility to get an econodist legend into a ggplot2 plot}
\usage{
add_econodist_legend(gg, legend, below = "subtitle", just = c("left",
  "center", "right"), legend_height = unit(16, "points"),
  spacer = unit(10, "points"))
}
\arguments{
\item{gg}{ggplot2 plot object to add}

\item{legend}{legend grob (any grob, really)}

\item{below}{which named gtable element to stick it below?}

\item{just}{legend position: "\code{left}" (aligned with left veritcal axis),
"\code{center}" (center of plot panel), or "\code{right}" (aligned with right
vertical axis). Uses \code{\link[=pmatch]{pmatch()}} so partial matching is supported.}

\item{legend_height}{height of the legend row}

\item{spacer}{height of the spacer that is put below \code{legend}?}
}
\description{
Helper utility to get an econodist legend into a ggplot2 plot
}
\seealso{
Other Econodist legend helpers: \code{\link{econodist_legend_grob}},
  \code{\link{left_align}}
}
\concept{Econodist legend helpers}
