% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggeconodist-package.R
\docType{package}
\name{ggeconodist}
\alias{ggeconodist}
\alias{ggeconodist-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/ggeconodist}
\item BugReports: \url{https://gitlab.com/hrbrmstr/ggeconodist/issues}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/ggeconodist}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/ggeconodist/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
